Divided transform into tools and descriptors
