+++ mlcolvar/core/transform/__init__.py
@@ -1,12 +1,16 @@
 __all__ = ["Transform","Normalization","Statistics","SwitchingFunctions","PairwiseDistances","RadiusGraph","EigsAdjMat","ContHist","RDF","Inverse",'TorsionalAngle']
 
 from .transform import *
-from .normalization import *
 from .utils import *
-from .eigs_adjacency_matrix import *
-from .pairwise_distances import *
+
+from .tools.normalization import *
+from .tools.switching_functions import *
+from .tools.continuous_hist import *
+from .tools.utils import *
+
+from .descriptors.eigs_adjacency_matrix import *
+from .descriptors.pairwise_distances import *
+from .descriptors.torsional_angle import *
 from .radius_graph import *
-from .switching_functions import *
-from .continuous_hist import *
 from .radial_distribution_function import *
-from .torsional_angle import *
+
