+++ mlcolvar/core/transform/descriptors/eigs_adjacency_matrix.py
@@ -1,7 +1,7 @@
 import torch
 
 from mlcolvar.core.transform import Transform
-from mlcolvar.core.transform.utils import compute_adjacency_matrix
+from mlcolvar.core.transform.descriptors.utils import compute_adjacency_matrix
 
 from typing import Union
 
@@ -79,7 +79,7 @@ class EigsAdjMat(Transform):
         return eigs
 
 def test_eigs_of_adj_matrix():
-    from mlcolvar.core.transform.switching_functions import SwitchingFunctions
+    from mlcolvar.core.transform.tools.switching_functions import SwitchingFunctions
     
     n_atoms=2
     pos = torch.Tensor([ [ [0., 0., 0.],
