+++ mlcolvar/core/transform/descriptors/pairwise_distances.py
@@ -1,7 +1,7 @@
 import torch
 
 from mlcolvar.core.transform import Transform
-from mlcolvar.core.transform.utils import compute_distances_matrix
+from mlcolvar.core.transform.descriptors.utils import compute_distances_matrix
 
 from typing import Union
