+++ mlcolvar/core/transform/radial_distribution_function.py
@@ -1,7 +1,8 @@
 import torch
 
-from mlcolvar.core.transform import Transform,PairwiseDistances
-from mlcolvar.core.transform.utils import easy_KDE
+from mlcolvar.core.transform import Transform
+from mlcolvar.core.transform.descriptors.pairwise_distances import PairwiseDistances
+from mlcolvar.core.transform.tools.utils import easy_KDE
 
 from typing import Union
