+++ mlcolvar/core/transform/tools/switching_functions.py
