+++ mlcolvar/tests/test_core_transform_adjacencymatrix.py
@@ -1,4 +1,4 @@
-from mlcolvar.core.transform.eigs_adjacency_matrix import test_eigs_of_adj_matrix
+from mlcolvar.core.transform.descriptors.eigs_adjacency_matrix import test_eigs_of_adj_matrix
 
 if __name__ == "__main__":
     test_eigs_of_adj_matrix()
