+++ mlcolvar/tests/test_core_transform_descriptors_utils.py
