+++ mlcolvar/tests/test_core_transform_normalization.py
@@ -1,6 +1,4 @@
-import pytest
-
-from mlcolvar.core.transform.normalization import test_normalization
+from mlcolvar.core.transform.tools.normalization import test_normalization
 
 if __name__ == "__main__":
     test_normalization()
