+++ mlcolvar/tests/test_core_transform_pairwisedistances.py
@@ -1,4 +1,4 @@
-from mlcolvar.core.transform.pairwise_distances import test_pairwise_distances
+from mlcolvar.core.transform.descriptors.pairwise_distances import test_pairwise_distances
 
 if __name__ == "__main__":
     test_pairwise_distances()
