+++ mlcolvar/tests/test_core_transform_switchingfunctions.py
@@ -1,4 +1,4 @@
-from mlcolvar.core.transform.switching_functions import test_switchingfunctions
+from mlcolvar.core.transform.tools.switching_functions import test_switchingfunctions
 
 if __name__ == "__main__":
     test_switchingfunctions()
