+++ mlcolvar/tests/test_core_transform_tools_utils.py
@@ -1,0 +1,4 @@
+from mlcolvar.core.transform.tools.utils import test_statistics
+
+if __name__ == "__main__":
+    test_statistics()
