+++ mlcolvar/tests/test_core_transform_torsionalangle.py
@@ -1,4 +1,4 @@
-from mlcolvar.core.transform.torsional_angle import test_torsional_angle
+from mlcolvar.core.transform.descriptors.torsional_angle import test_torsional_angle
 
 if __name__ == "__main__":
     test_torsional_angle()
